告警整改&&补充tdd

+++ rosen/modules/hyper_graphic_manager/core/frame_rate_manager/hgm_frame_rate_manager.cpp
@@ -409,7 +409,7 @@ void HgmFrameRateManager::UniProcessDataForLtpo(uint64_t timestamp,
             auto expectedRange = linker.second->GetExpectedRange();
             HgmEnergyConsumptionPolicy::Instance().GetUiIdleFps(expectedRange);
             if ((expectedRange.type_ & ANIMATION_STATE_FIRST_FRAME) != 0 &&
-                expectedRange.preferred_ < currRefreshRate_) {
+                expectedRange.preferred_ < static_cast<int32_t>(currRefreshRate_)) {
                 expectedRange.Set(currRefreshRate_, currRefreshRate_, currRefreshRate_);
             }
             finalRange.Merge(expectedRange);
